Add static files to zip
